refactor: inline base64 utilities into rspack_util
