--- crates/rspack_base64/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-description       = "rspack base64"
-edition.workspace = true
-license           = "MIT"
-name              = "rspack_base64"
-repository        = "https://github.com/web-infra-dev/rspack"
-version.workspace = true
-
-[dependencies]
-base64-simd = { workspace = true }
-regex       = { workspace = true }
