+++ crates/rspack_core/Cargo.toml
@@ -32,7 +32,6 @@ rayon = { workspace = true }
 regex = { workspace = true }
 rkyv = { workspace = true, optional = true }
 ropey = { workspace = true }
-rspack_base64 = { workspace = true }
 rspack_cacheable = { workspace = true }
 rspack_collections = { workspace = true }
 rspack_dojang = { workspace = true }
