+++ crates/rspack_plugin_asset/Cargo.toml
@@ -11,7 +11,6 @@ version.workspace = true
 async-trait      = { workspace = true }
 mime_guess       = { workspace = true }
 rayon            = { workspace = true }
-rspack_base64    = { workspace = true }
 rspack_cacheable = { workspace = true }
 rspack_core      = { workspace = true }
 rspack_error     = { workspace = true }
