+++ crates/rspack_plugin_html/Cargo.toml
@@ -16,7 +16,6 @@ futures           = { workspace = true }
 itertools         = { workspace = true }
 path-clean        = { workspace = true }
 rayon             = { workspace = true }
-rspack_base64     = { workspace = true }
 rspack_core       = { workspace = true }
 rspack_dojang     = { workspace = true }
 rspack_error      = { workspace = true }
