+++ crates/rspack_plugin_schemes/src/data_uri.rs
@@ -7,6 +7,7 @@ use rspack_core::{
 };
 use rspack_error::Result;
 use rspack_hook::{plugin, plugin_hook};
+use rspack_util::base64;
 
 static URI_REGEX: LazyLock<Regex> = LazyLock::new(|| {
   Regex::new(r"(?i)^data:([^;,]+)?((?:;[^;,]+)*?)(?:;(base64)?)?,(.*)$").expect("Invalid Regex")
@@ -62,8 +63,8 @@ async fn read_resource(&self, resource_data: &ResourceData) -> Result<Option<Con
   {
     let body = captures.get(4).expect("should have data uri body").as_str();
     let is_base64 = captures.get(3).is_some();
-    if is_base64 && let Some(cleaned) = rspack_base64::clean_base64(body) {
-      return match rspack_base64::decode_to_vec(cleaned.as_bytes()) {
+    if is_base64 && let Some(cleaned) = base64::clean_base64(body) {
+      return match base64::decode_to_vec(cleaned.as_bytes()) {
         Ok(buffer) => Ok(Some(Content::Buffer(buffer))),
         Err(_) => Ok(Some(Content::String(resource_data.resource().to_string()))),
       };
