+++ crates/rspack_plugin_sri/Cargo.toml
@@ -16,7 +16,6 @@ indexmap                        = { workspace = true }
 pathdiff                        = { workspace = true }
 rayon                           = { workspace = true }
 regex                           = { workspace = true }
-rspack_base64                   = { workspace = true }
 rspack_cacheable                = { workspace = true }
 rspack_collections              = { workspace = true }
 rspack_core                     = { workspace = true }
