+++ crates/rspack_plugin_sri/src/integrity.rs
@@ -1,4 +1,5 @@
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
+use rspack_util::base64;
 use sha2::{Digest, Sha256, Sha384, Sha512};
 
 // https://www.w3.org/TR/2016/REC-SRI-20160623/#cryptographic-hash-functions
@@ -38,19 +39,19 @@ fn create_hash(hash_func: &SubresourceIntegrityHashFunction, source: &str) -> St
       let mut hasher = Sha256::new();
       hasher.update(source);
       let digest = &hasher.finalize()[..];
-      format!("sha256-{}", rspack_base64::encode_to_string(digest))
+      format!("sha256-{}", base64::encode_to_string(digest))
     }
     SubresourceIntegrityHashFunction::Sha384 => {
       let mut hasher = Sha384::new();
       hasher.update(source);
       let digest = &hasher.finalize()[..];
-      format!("sha384-{}", rspack_base64::encode_to_string(digest))
+      format!("sha384-{}", base64::encode_to_string(digest))
     }
     SubresourceIntegrityHashFunction::Sha512 => {
       let mut hasher = Sha512::new();
       hasher.update(source);
       let digest = &hasher.finalize()[..];
-      format!("sha512-{}", rspack_base64::encode_to_string(digest))
+      format!("sha512-{}", base64::encode_to_string(digest))
     }
   }
 }
