+++ crates/rspack_util/Cargo.toml
@@ -9,6 +9,7 @@ version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+base64-simd      = { workspace = true }
 bitflags         = { workspace = true }
 concat-string    = { workspace = true }
 cow-utils        = { workspace = true }
